feat: 헤더컴포넌트 서버인증 후 UI구현#14
