+++ components/shared/header/HeaderLayout.tsx
@@ -1,7 +1,25 @@
 'use client';
 
 import HeaderUi from './Ui';
+import { userApi } from '@/service/auth/api';
 
-export default function HeaderLayout() {
-  return <HeaderUi />;
+interface HeaderLayoutProps {
+  isAuthenticated: boolean;
+}
+
+export default function HeaderLayout({ isAuthenticated }: HeaderLayoutProps) {
+  const handleLogin = () => {
+    window.location.href = userApi.getKakaoLogin();
+  };
+
+  const handleLogout = async () => {
+    try {
+      await userApi.logout();
+      window.location.href = '/';
+    } catch (error) {
+      console.error('Logout failed:', error);
+    }
+  };
+
+  return <HeaderUi isAuthenticated={isAuthenticated} onLogin={handleLogin} onLogout={handleLogout} />;
 }
